% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC2253.R
\name{format.basedn.class}
\alias{format.basedn.class}
\title{format.basedn.class}
\usage{
\method{format}{basedn.class}(x, ...)
}
\arguments{
\item{x}{a 'ldapquery' object}

\item{...}{ignored}
}
\description{
format.basedn.class
}
