% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.R
\name{format.ip.class}
\alias{format.ip.class}
\title{format.ip.class}
\usage{
\method{format}{ip.class}(x, ...)
}
\arguments{
\item{x}{an 'ip.class' object}

\item{...}{ignored}
}
\description{
format.ip.class
}
