% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.R
\name{format.ipv4.class}
\alias{format.ipv4.class}
\title{format.ipv4.class}
\usage{
\method{format}{ipv4.class}(x, ...)
}
\arguments{
\item{x}{an 'ipv4.class' object}

\item{...}{ignored}
}
\description{
format.ipv4.class
}
