% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.R
\name{format.ipv4.subnet}
\alias{format.ipv4.subnet}
\title{format.ipv4.subnet}
\usage{
\method{format}{ipv4.subnet}(x, ...)
}
\arguments{
\item{x}{an 'ipv4.subnet' object}

\item{...}{ignored}
}
\description{
format.ipv4.subnet
}
