% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC2253.R
\name{format.ldapkv}
\alias{format.ldapkv}
\title{format.ldapkv}
\usage{
\method{format}{ldapkv}(x, collapse = ": ", ...)
}
\arguments{
\item{x}{ldapkv object}

\item{sep}{the 'character' used to divide key and value}
}
\description{
format.ldapkv
}
