% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC2253.R
\name{format.ldapquery}
\alias{format.ldapquery}
\title{format.ldapquery}
\usage{
\method{format}{ldapquery}(x, ...)
}
\arguments{
\item{x}{a 'ldapquery' object}

\item{...}{ignored}
}
\description{
format.ldapquery
}
