% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.R
\name{ipv4.subnet.broadcast}
\alias{ipv4.subnet.broadcast}
\title{ipv4.subnet.broadcast}
\usage{
ipv4.subnet.broadcast(subnet)
}
\arguments{
\item{subnet}{an 'ipv4.subnet' object}
}
\description{
ipv4.subnet.broadcast
}
