% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.subnet.R
\name{ipv4.subnet}
\alias{ipv4.subnet}
\title{ipv4.subnet}
\usage{
ipv4.subnet(x, ...)
}
\arguments{
\item{'...'}{5 'numeric' or 'integer' type subnet identifiers}
}
\description{
a container for a valid RFC1918 IPv4 subnet
}
