% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.subnetlist.R
\name{ipv4.subnetlist}
\alias{ipv4.subnetlist}
\title{ipv4.subnetlist}
\usage{
ipv4.subnetlist(x, ...)
}
\arguments{
\item{x}{a 'character' filename like '/etc/networks' 
or a 'character' vector like scan('/etc/networks',what='character',sep='\\n')
or a 'list' of 'ipv4.subnet' objects}
}
\description{
ipv4.subnetlist
}
