% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.ipv4list.R
\name{ipv4list}
\alias{ipv4list}
\title{ipv4list}
\usage{
ipv4list(x, ...)
}
\arguments{
\item{x}{a 'character' filename like '/etc/hosts' 
or a 'character' vector like scan('/etc/hosts',what='character',sep='\\n')
or a 'list' of 'ipv4' objects}
}
\description{
ipv4list
}
