% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC2253.R
\name{ldapkv}
\alias{ldapkv}
\title{ldapkv}
\usage{
ldapkv(key, value)
}
\arguments{
\item{key}{a valid 'RFC2253string'}

\item{value}{a valid 'RFC2253string'}
}
\description{
a container for a valid LDAP key/value pair
}
