% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC2253.R
\name{ldapquerylist}
\alias{ldapquerylist}
\title{ldapquerylist}
\usage{
ldapquerylist(...)
}
\arguments{
\item{...}{either 1 'list' of 'ldapquery' objects, or 
arbitrary 'ldapquery' objects}
}
\description{
ldapquerylist
}
