% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.R
\name{<.ipv4.class}
\alias{<.ipv4.class}
\title{>.ipv4.class}
\usage{
\method{<}{ipv4.class}(e1, e2)
}
\arguments{
\item{e1}{an 'ipv4.class' object}

\item{e2}{an 'ipv4.class' object}
}
\description{
>.ipv4.class
}
