% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mac802.R
\name{mac802list}
\alias{mac802list}
\title{mac802list}
\usage{
mac802list(x, ...)
}
\arguments{
\item{x}{a 'character' filename like '/etc/ethers' 
or a 'character' vector like scan('/etc/ethers',what='character',sep='\\n')
or a 'list' of 'mac802' objects}
}
\description{
mac802list
}
