% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.R
\name{+.ip.class}
\alias{+.ip.class}
\title{+.ip.class}
\usage{
\method{+}{ip.class}(e1, e2)
}
\arguments{
\item{e1}{an 'ip.class' object}

\item{e2}{an 'ip.class' object or length==1 integer}
}
\description{
+.ip.class
}
