% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC2253.R
\name{+.ldapkvlist}
\alias{+.ldapkvlist}
\title{+.ldapkvlist}
\usage{
"+.ldapkvlist"(e1, e2)
}
\arguments{
\item{e1}{a 'ldapkvlist' object}

\item{e2}{a 'ldapkvlist' or 'ldapkv' object}
}
\description{
+.ldapkvlist
}
