% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC2253.R
\name{+.ldapquerylist}
\alias{+.ldapquerylist}
\title{+.ldapquerylist}
\usage{
"+.ldapquerylist"(e1, e2)
}
\arguments{
\item{e1}{a 'ldapquerylist' object}

\item{e2}{a 'ldapquerylist' or 'ldapquery' object}
}
\description{
+.ldapquerylist
}
