% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.R
\name{print.ipv4.subnet}
\alias{print.ipv4.subnet}
\title{print.ipv4.subnet}
\usage{
\method{print}{ipv4.subnet}(x, ...)
}
\arguments{
\item{x}{an 'ipv4.subnet' object}

\item{...}{passed to print.character}
}
\description{
print.ipv4.subnet
}
