% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mac802.R
\name{print.mac802}
\alias{print.mac802}
\title{print.mac802}
\usage{
\method{print}{mac802}(x, ...)
}
\arguments{
\item{x}{an 'mac802' object}

\item{...}{passed to print.character}
}
\description{
print.mac802
}
