% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1123.R
\docType{data}
\name{RFC1123HostnameRegex}
\alias{RFC1123HostnameRegex}
\title{RFC1123HostnameRegex}
\format{An object of class \code{character} of length 1.}
\usage{
RFC1123HostnameRegex
}
\description{
A regex to identify valid RFC1123 RFC1123.host characters
}
\keyword{datasets}
