% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RFC2253Regex}
\alias{RFC2253Regex}
\title{RFC2253Regex}
\format{An object of class \code{character} of length 1.}
\usage{
RFC2253Regex
}
\description{
A regex to identify valid RFC2253 string characters
}
\keyword{datasets}
