% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RFC2253SpecialRegex}
\alias{RFC2253SpecialRegex}
\title{RFC2253SpecialRegex}
\format{An object of class \code{character} of length 1.}
\usage{
RFC2253SpecialRegex
}
\description{
A regex to identify valid RFC2253 special characters
}
\keyword{datasets}
