% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1123.R
\name{valid.hostname.class}
\alias{valid.hostname.class}
\title{valid.hostname.class}
\usage{
valid.hostname.class(x)
}
\arguments{
\item{x}{test object}
}
\description{
valid.hostname.class
}
\examples{
hexchars <-'123456789ABCDEF'
stopifnot(valid.hostname.class(vector())==FALSE)
stopifnot(valid.hostname.class('')==FALSE)
stopifnot(valid.hostname.class(paste(collapse='',rep(4,hexchars)))==FALSE)
stopifnot(valid.hostname.class(':')==FALSE)
stopifnot(valid.hostname.class(paste(collapse='',c(letters,LETTERS,'-')))==TRUE)
}
