% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1123.R
\name{hostname.class}
\alias{hostname.class}
\title{hostname.class}
\usage{
hostname.class(hostname)
}
\arguments{
\item{hostname}{a 'character' identifier}
}
\description{
a container for a valid RFC1123 hostname
}
