% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1918.R
\name{ipv4.subnetlist}
\alias{ipv4.subnetlist}
\title{ipv4.subnetlist}
\usage{
ipv4.subnetlist(x, ...)
}
\arguments{
\item{x}{a 'character' filename like '/etc/networks' 
or a character vector like scan('/etc/networks',what='character',sep='\\n')}
}
\description{
ipv4.subnetlist
}
