% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC2253.R
\name{ldapquery}
\alias{ldapquery}
\title{ldapquery}
\usage{
ldapquery(pkey, basedn, skeylist = list(), kvlist = list())
}
\arguments{
\item{pkey}{a 'ldapkv' the primary key for the query}

\item{basedn}{a valid 'domain.class' or a 'basedn.class' the base dn for the query}

\item{skeylist}{a 'list' of 'ldapkv' objects (default==0) the secondary keys for the query}

\item{kvlist}{a 'list' of 'ldapkv' objects (default==0) the content of the query}
}
\description{
a container for a valid LDAP query
}
