% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFC1123.R
\name{valid.domain.class}
\alias{valid.domain.class}
\title{valid.domain.class}
\usage{
valid.domain.class(x)
}
\arguments{
\item{x}{test object}
}
\description{
valid.domain.class
}
\examples{
hexchars <-'0123456789ABCDEF'
stopifnot(valid.domain.class(vector())==FALSE)
stopifnot(valid.domain.class('')==FALSE)
stopifnot(valid.domain.class(paste(collapse='',rep(16,hexchars)))==FALSE)
stopifnot(valid.domain.class(':')==FALSE)
stopifnot(valid.domain.class(paste(collapse='',c(letters,,'.',LETTERS,'-')))==TRUE)
}
